refactor: split types/general.ts into focused shared modules
DateRange, MoneyFormat + Direction, ReportingBasis, S3PresignedUrl and View each get their own
module under types/shared. LoadedStatus and BaseSelectOption had one consumer apiece and now
live there.
