+++ src/components/features/cashflowStatement/StatementOfCashFlow/StatementOfCashFlow.tsx
@@ -1,8 +1,8 @@
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { type View as ViewType } from '@internal-types/general'
 import type { TimeRangePickerConfig } from '@internal-types/reports'
+import { type View as ViewType } from '@internal-types/shared/viewport'
 import { useGetStatementOfCashFlow } from '@api/businesses/[business-id]/reports/cashflow-statement/get'
 import { useReportsCompactHeader } from '@hooks/features/reports/useReportsCompactHeader'
 import { useResolvedReportView } from '@hooks/features/reports/useResolvedReportView'
