+++ src/components/features/generalLedger/ParentAccountComboBox/useParentAccountOptions.ts
@@ -1,8 +1,12 @@
 import { useMemo } from 'react'
 
-import { type BaseSelectOption } from '@internal-types/general'
 import { type LedgerBalancesSchemaType } from '@schemas/generalLedger/ledgerBalances'
 import { flattenAccounts } from '@features/generalLedger/utils'
+
+type BaseSelectOption = {
+  label: string
+  value: string | number
+}
 
 const isAlphanumeric = (char: string) => /[\p{L}\p{N}]/u.test(char)
