+++ src/components/features/profitAndLoss/ProfitAndLossDownloadButton/ProfitAndLossFullReportDownloadButton.tsx
@@ -1,7 +1,7 @@
 import { useContext, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 
-import { type MoneyFormat } from '@internal-types/general'
+import { type MoneyFormat } from '@internal-types/shared/money'
 import { getProfitAndLossExcel } from '@hooks/legacy/useDownloadProfitAndLoss'
 import { useBuildKeyInputs } from '@hooks/utils/swr/useBuildKeyInputs'
 import { ProfitAndLossContext } from '@providers/profitAndLoss/ProfitAndLossContext/ProfitAndLossContext'
