+++ src/components/ui/DatePickers/DateCalendar/DateCalendar.tsx
@@ -1,7 +1,7 @@
 import { type DateValue } from '@internationalized/date'
 import { ChevronLeft, ChevronRight } from 'lucide-react'
 
-import type { View } from '@internal-types/general'
+import type { View } from '@internal-types/shared/viewport'
 import { Button } from '@ui/Button/Button'
 import { Calendar, CalendarCell, CalendarGrid, CalendarGridBody, CalendarGridHeader, CalendarHeaderCell } from '@ui/Calendar/Calendar'
 import { HStack } from '@ui/Stack/Stack'
