+++ src/hooks/api/businesses/[business-id]/reports/profit-and-loss/lines/get.ts
@@ -1,4 +1,4 @@
-import { type ReportingBasis } from '@internal-types/general'
+import { type ReportingBasis } from '@internal-types/shared/reporting'
 import { UnwrappedDataResponseSchema } from '@schemas/common/utils'
 import { type PnlDetailLineSchema, PnlDetailLinesDataSchema } from '@schemas/profitAndLoss/pnlDetailLines'
 import { get } from '@utils/shared/api/authenticatedHttp'
