+++ src/hooks/api/businesses/[business-id]/reports/transactions/exports/excel/get.ts
@@ -1,4 +1,4 @@
-import type { S3PresignedUrl } from '@internal-types/general'
+import type { S3PresignedUrl } from '@internal-types/shared/s3'
 import { getAsMutation } from '@utils/shared/api/getAsMutation'
 import { getWithQuery } from '@utils/shared/api/getWithQuery'
 import type { UseBankTransactionsOptions } from '@api/businesses/[business-id]/bank-transactions/get'
