+++ src/hooks/features/profitAndLoss/useProfitAndLoss.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useMemo, useState } from 'react'
 
-import { type ReportingBasis } from '@internal-types/general'
+import { type ReportingBasis } from '@internal-types/shared/reporting'
 import { getNextSortOrder, SortOrder, type SortParams } from '@internal-types/utility/pagination'
 import {
   applyShare,
