+++ src/hooks/features/profitAndLoss/useProfitAndLossLTM.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useMemo, useRef, useState } from 'react'
 
-import { type ReportingBasis } from '@internal-types/general'
+import { type ReportingBasis } from '@internal-types/shared/reporting'
 import type { ProfitAndLossSummary } from '@schemas/profitAndLoss/profitAndLossSummaries'
 import { useGetProfitAndLossSummaries } from '@api/businesses/[business-id]/reports/profit-and-loss-summaries/get'
 import type { ChartWindow } from '@features/profitAndLoss/ProfitAndLossChart/getChartWindow'
