+++ src/hooks/utils/size/useElementViewSize.tsx
@@ -1,6 +1,6 @@
 import { useLayoutEffect, useRef, useState } from 'react'
 
-import { type View } from '@internal-types/general'
+import { type View } from '@internal-types/shared/viewport'
 import { BREAKPOINTS } from '@utils/shared/size/screenSizeBreakpoints'
 
 export const useElementViewSize = <T extends HTMLElement>() => {
