+++ src/index.tsx
@@ -61,5 +61,5 @@ export { useLayerContext } from './providers/global/LayerContext/LayerContext'
 export { type LayerEvent, LayerEventComponent, LayerEventType } from './providers/global/LayerProvider/layerEvents'
 export { type EventCallbacks, LayerProvider } from './providers/global/LayerProvider/LayerProvider'
 export { DisplayState } from './types/bankTransactions'
-export { Direction } from './types/general'
+export { Direction } from './types/shared/money'
 export { SupportedLocale } from './utils/shared/i18n/supportedLocale'
