+++ src/msw/api/businesses/[business-id]/bank-transactions/[bank-transaction-id]/documents/store.ts
@@ -1,4 +1,4 @@
-import { type S3PresignedUrl } from '@internal-types/general'
+import { type S3PresignedUrl } from '@internal-types/shared/s3'
 
 import { createMockStore } from '@msw/utils/createMockStore'
