+++ src/msw/utils/csvPresignedUrl.ts
@@ -1,4 +1,4 @@
-import { type S3PresignedUrl } from '@internal-types/general'
+import { type S3PresignedUrl } from '@internal-types/shared/s3'
 
 export const formatCsvCents = (cents: number) => (cents / 100).toFixed(2)
