+++ src/types/layerContext.ts
@@ -1,4 +1,4 @@
-import { type DateRange } from '@internal-types/general'
+import { type DateRange } from '@internal-types/shared/dateRange'
 import { type ToastData } from '@internal-types/toast'
 import { type AccountingConfigurationSchemaType } from '@schemas/business/accountingConfiguration'
 import { type Business } from '@schemas/business/business'
