+++ src/types/shared/dateRange.ts
@@ -1,0 +1,4 @@
+export type DateRange<T = Date> = {
+  startDate: T
+  endDate: T
+}
