+++ src/types/shared/money.ts
@@ -1,0 +1,6 @@
+export type MoneyFormat = 'CENTS' | 'DOLLAR_STRING'
+
+export enum Direction {
+  CREDIT = 'CREDIT',
+  DEBIT = 'DEBIT',
+}
