+++ src/types/shared/reporting.ts
@@ -1,0 +1,4 @@
+import type { EnumWithUnknownValues } from '@internal-types/utility/enumWithUnknownValues'
+
+type StrictReportingBasis = 'CASH' | 'CASH_COLLECTED' | 'ACCRUAL'
+export type ReportingBasis = EnumWithUnknownValues<StrictReportingBasis>
