+++ src/types/shared/viewport.ts
@@ -1,0 +1,1 @@
+export type View = 'mobile' | 'tablet' | 'desktop'
